// client/src/types/vehicular.types.ts
export interface Vehicular_types {
  id?: number;
  fecha: string;
  nombre: string;
  cedula: string;
  placa: string;
  kilometraje: string;
  l_baja: string;
  l_alta: string;
  l_marcha_atras: string;
  l_interior: string;
  l_freno: string;
  estacionarias: string;
  viraje_derecho: string;
  viraje_izquierdo: string;
  observacion1?: string;
  radio: string;
  pitos: string;
  parabrisas: string;
  vidrios_laterales: string;
  manillas_alza_vidrios: string;
  Bateria: string;
  espejos_retrovisores: string;
  cerraduras: string;
  plumillas_limpia_vidrios: string;
  dispositivo_velocidad: string;
  tapa_aceite: string;
  tapa_gasolina: string;
  tapa_radiador: string;
  tapetes: string;
  varilla_aceite: string;
  revision_externa?: string;
  observacion2?: string;
  delanteros: string;
  traseros: string;
  repuestos: string;
  observacion3?: string;
  principal: string;
  emergencia: string;
  observacion4?: string;
  extintor_vigente: string;
  cinturones_seguridad: string;
  caja_herramientas: string;
  gato: string;
  cuna: string;
  llave_rueda: string;
  botiquin: string;
  triangulo: string;
  observacion5?: string;
  licencia_transito: string;
  seguro: string;
  licencia_conduccion: string;
  revision: string;
  observacion6?: string;
  liquido_freno: string;
  acite: string;
  refrigerante: string;
  observacion7?: string;
  amortiguadores: string;
  cardan: string;
  caja_cambios: string;
  observacion8?: string;
  imagen_inspeccion1?: string;
  observacion9?: string;
  imagen_inspeccion2?: string;
  observacion10?: string;
  imagen_inspeccion3?: string;
  observacion11?: string;
  firma_administracion?: string;
  firma_conductor?: string;
}